decoupled dialog and inventory system
