+++ systems/unitinventorysystem.go
@@ -2,9 +2,9 @@ package systems
 
 import (
 	"engo.io/ecs"
+	"github.com/MrTrustworthy/fargo/events"
+	"github.com/MrTrustworthy/fargo/ui"
 )
-
-
 
 type UnitInventorySystem struct {
 	*ecs.World
@@ -12,7 +12,23 @@ type UnitInventorySystem struct {
 
 func (is *UnitInventorySystem) New(world *ecs.World) {
 	is.World = world
+	events.Mailbox.Listen(events.INVENTORY_SHOW_EVENT, is.getHandleShowInventory())
+
 }
 
-func (is *UnitInventorySystem) Update(dt float32) {}
+func (is *UnitInventorySystem) getHandleShowInventory() func(msg events.BaseEvent) {
+	return func(msg events.BaseEvent) {
+		simsg, ok := msg.(events.ShowInventoryEvent)
+		if !ok || simsg.Unit == nil {
+			return
+		}
+		inventoryDialog := ui.NewInventoryDialog()
+		events.Mailbox.Dispatch(events.DialogShowEvent{Dialog: inventoryDialog})
+
+	}
+}
+
+
+
+func (is *UnitInventorySystem) Update(dt float32)        {}
 func (is *UnitInventorySystem) Remove(e ecs.BasicEntity) {}
